requirements.txt
torch
transformers
datasets
scikit-learn
pandas
numpy
matplotlib
seaborn
tqdm
sentence-transformers
umap-learn
jupyter
ipykernel
wandb
accelerate
openai
umap-learn
python-dotenv
